refactor(iroh): move rpc handling into its own module
